refactor(config): point getExperimentalFlag callers at experimentalConfigStore
getExperimentalFlag now lives in @/config/experimentalConfigStore. Update
the five non-test callers and their jest mocks to import from the source
module instead of routing through experimentalHooks's re-export.

experimentalHooks no longer re-exports getExperimentalFlag; the
`export *` of constants is preserved temporarily and removed in a
follow-up commit.
