+++ src/features/positions/stores/transform/locked.test.ts
@@ -6,9 +6,8 @@ import { DetailType, PositionName } from '../../types/generated/positions/positi
 import { usePositionsStore } from '../positionsStore';
 
 // Mock config to avoid React Native gesture handler imports
-jest.mock('@/config/experimentalHooks', () => ({
+jest.mock('@/config/experimentalConfigStore', () => ({
   getExperimentalFlag: jest.fn(() => false),
-  DEFI_POSITIONS_THRESHOLD_FILTER: 'defi_positions_threshold_filter',
 }));
 
 jest.mock('@/state/backendNetworks/backendNetworks', () => ({
